src/main/kotlin/uk/gov/justice/digital/hmpps/prisonperson/service/PhysicalAttributesSyncService.kt
package uk.gov.justice.digital.hmpps.prisonperson.service

import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional
import uk.gov.justice.digital.hmpps.prisonperson.dto.PhysicalAttributesDto
import uk.gov.justice.digital.hmpps.prisonperson.dto.PhysicalAttributesSyncRequest

@Service
@Transactional
class PhysicalAttributesSyncService() {
  fun sync(
    prisonerNumber: String,
    request: PhysicalAttributesSyncRequest,
  ): PhysicalAttributesDto {
    TODO("CDPS-776: Update the physical attributes and history table")
  }
}
